CEP-7: SAI source package

+++ src/java/org/apache/cassandra/io/util/SequentialWriter.java
@@ -168,7 +168,7 @@ public class SequentialWriter extends BufferedDataOutputStreamPlus implements Tr
         this.option = option;
     }
 
-    public void skipBytes(int numBytes) throws IOException
+    public void skipBytes(long numBytes) throws IOException
     {
         flush();
         fchannel.position(fchannel.position() + numBytes);
